Feature/4 spec update (#446)
* handled a bunch of todo's and the latest spec update
* fixed serialization to properly ignore default values for [Optional] properties
* Added tests to check for null

+++ src/Protocol/Features/Document/FoldingRangeFeature.cs
@@ -45,39 +45,35 @@ namespace OmniSharp.Extensions.LanguageServer.Protocol
             /// The zero-based line number from where the folded range starts.
             /// </summary>
             /// <remarks>
-            /// TODO: UPDATE THIS next version
             /// <see cref="uint"/> in the LSP spec
             /// </remarks>
-            public long StartLine { get; init; }
+            public int StartLine { get; init; }
 
             /// <summary>
             /// The zero-based character offset from where the folded range starts. If not defined, defaults to the length of the start line.
             /// </summary>
             /// <remarks>
-            /// TODO: UPDATE THIS next version
             /// <see cref="uint"/> in the LSP spec
             /// </remarks>
             [Optional]
-            public long? StartCharacter { get; init; }
+            public int? StartCharacter { get; init; }
 
             /// <summary>
             /// The zero-based line number where the folded range ends.
             /// </summary>
             /// <remarks>
-            /// TODO: UPDATE THIS next version
             /// <see cref="uint"/> in the LSP spec
             /// </remarks>
-            public long EndLine { get; init; }
+            public int EndLine { get; init; }
 
             /// <summary>
             /// The zero-based character offset before the folded range ends. If not defined, defaults to the length of the end line.
             /// </summary>
             /// <remarks>
-            /// TODO: UPDATE THIS next version
             /// <see cref="uint"/> in the LSP spec
             /// </remarks>
             [Optional]
-            public long? EndCharacter { get; init; }
+            public int? EndCharacter { get; init; }
 
             /// <summary>
             /// Describes the kind of the folding range such as `comment' or 'region'. The kind
