+++ src/Protocol/Server/WorkDone/IWorkDoneObserver.cs
@@ -6,6 +6,6 @@ namespace OmniSharp.Extensions.LanguageServer.Protocol.Server.WorkDone
     public interface IWorkDoneObserver : IObserver<WorkDoneProgress>, IDisposable
     {
         ProgressToken WorkDoneToken { get; }
-        void OnNext(string message, double? percentage, bool? cancellable);
+        void OnNext(string message, int? percentage, bool? cancellable);
     }
 }
