+++ src/Protocol/Server/WorkDone/WorkDoneObserver.cs
@@ -50,7 +50,7 @@ namespace OmniSharp.Extensions.LanguageServer.Protocol.Server.WorkDone
 
         public ProgressToken WorkDoneToken { get; }
 
-        public void OnNext(string message, double? percentage, bool? cancellable) =>
+        public void OnNext(string message, int? percentage, bool? cancellable) =>
             OnNext(
                 new WorkDoneProgressReport {
                     Cancellable = cancellable ?? false,
